refactor(translator): move into src layout and use uv properly
