+++ pyproject.toml
@@ -150,4 +150,4 @@ django_settings_module = "config.settings.test"
 
 
 [tool.uv.workspace]
-exclude = ["scheduler"]
+exclude = ["scheduler", "translator"]
