+++ translator/pyproject.toml
@@ -1,0 +1,21 @@
+[project]
+name = "translator"
+dependencies = [
+  "aiohttp-sse-client>=0.2.1",
+  "behave>=1.2.6",
+  "coverage>=5.5",
+  "grpcio-tools>=1.69.0",
+  "websockets>=10.3",
+]
+requires-python = ">=3.13"
+version = "0.1.0"
+
+[dependency-groups]
+dev = [
+  "debugpy",
+  "pydevd-pycharm",
+]
+
+[build-system]
+requires = ["setuptools", "wheel"]
+build-backend = "setuptools.build_meta"
