--- translator/requirements/local.txt
@@ -1,5 +1,0 @@
--r base.txt
-
-debugpy
-# Pycharm might force you to be on the same version IDE as library, use caution.
-pydevd-pycharm
