+++ translator/src/translator/__init__.py
@@ -1,0 +1,1 @@
+"""Translator package for SCRAM."""
