+++ translator/src/translator/exceptions.py
